Bugfix: Move the duplicate subscriptions check
